OpenSSL Security Provider
This change adds support for the OpenSSL Security Provider which reads OpenSSL
artifacts (certificates and private keys) and configures Java truststores and
keystores with their contents.  As part of this addition, a significant change
was made to the JRE contribution to ensure that certain layers were also
marked with build (when they weren't previously) if the contents of those
layers made sense during build time use of the JRE (e.g jvmkill,
link-local-dns).

Signed-off-by: Ben Hale <bhale@pivotal.io>