Update JDK Certificates
Previously the addition of the OpenSSL Security Provider allowed OpenSSL
certificates (including contributed from outside the container) to be to be
used at runtime.  However, the strategy for doing this at build time couldn't
be the same.  This change modifies the JDK layer contributor so that those
certificates are updated at build time, before the JDK is used by other
buildpacks.

This change also modifies the Security Provider Configurer layer so that the
profile script is aware of whether it's running against a JRE or JDK.

Signed-off-by: Ben Hale <bhale@pivotal.io>