Accelerate JDK CA Certificate Loading
Previously the implementation of CA Certificate loading used successive calls
to the keytool.  This was horrifically slow, potentially resulting in minutes
to load 133 certificates.  This change updates the loading to read the
keystore into memory, load all certificates and then write the keystore at the
end, resulting in something more or less instant.

Signed-off-by: Ben Hale <bhale@vmware.com>