JRE CA Certificate Loading
Previously, container CA certificate loading for JREs at runtime was achieved
by loading a security provider that would create a delegating in-memory
collection of trusted CA certificates.  While this was fast and reliable, it
was difficult to explain to users and could have unintended side-effects.
With the creation of a Go-based helper for loading certificates into a JDK's
CA truststore, it became obvious that we should just do the same kind of
behavior at launch for JRE's as well.

This change removes the OpenSSL Security Provider and replaces it with the
openssl-certificate-loader helper application which loads the certificates
into the JRE's CA truststore more or less instantly on startup.

Signed-off-by: Ben Hale <bhale@vmware.com>