Missing limit_in_bytes file
Previously, the memory calculator script left in built images assumed that the
images would always be run in environments with a
/sys/fs/cgroup/memory/memory.limit_in_byte file.  Unfortunately, some
environments (looking at you vSphere 7), don't have this file.  This change
adds a test for the existence of the file and if it does not exist, assumes
that the value is the same as if it was unset.

[resolves #5]

Signed-off-by: Ben Hale <bhale@vmware.com>