Load Certificates from $SSL_CERT_DIR
Previously, when the the CertificateLoader loaded, it would parse a single
magical file for certificates.  To support the customization of CA
certificates though, it should have been honoring the $SSL_CERT_DIR
path-delimited environment variable for the certificate locations.

This change adds support for parsing the contents of this environment
variable, falling back to the Ubuntu default when it is not set.

Signed-off-by: Ben Hale <bhale@vmware.com>