Add available memory fallback for memory calculator
This adds a new step where the memory calculation is done based on an estimation of how much memory is available for starting new applications, without swapping.

Decision order for the amount of memory used for calculation:
1. configured container memory limit
2. systems available memory
3. 1G fallback

This results in a better memory usage for cases where the containers memory limit was not set or can not be set.

See https://man7.org/linux/man-pages/man5/proc.5.html
Ref https://github.com/paketo-buildpacks/java/issues/82