Ensure only one JRE layer is ever added
This resolves #75, which is an edge case hit when a JVM provider does not have a JRE. In that case, the buildpack would have installed the JDK, then installed a JRE, which would have been the fallback JDK. The net result is two layers that are the same, both with a JDK. However, if you try to use that with Maven to build from source, you'd get a failure because the second JDK layer overwrites the layer metadata causing the JDK to not be available at build time.

This PR resolves the issue by ensuring that the JDK layer is not installed if a JDK is going to be used as a fallback JRE. In that case, it'll skip the JDK layer and add the JRE layer with the JDK as the JRE. It then ensure that metadata flags are set to launch=true, build=true and cache=true so this single layer works for both building and running the application.