Changes to heap dump path:
- Pass in a file, not a directory
- Do not use a UUID, instead use an RFC3339 formatted date

The date is easier to correlate than a UUID, doesn't require a dependency, and should be unique enough. If we need something that's more unique, we can revisit this down the road.

In addition, we switch from passing a folder as the heap path, because each time a container runs it will at most only ever write one file. Passing the full path to a file instead of a directory allows us to put the date into the file name itself, which seems a little more useful.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>