Adds BPL_JVM_CLASS_ADJUSTMENT variable
This commit adds the BPL_JVM_CLASS_ADJUSTMENT environment variable. This
variable allows a developer (usually a buildpack developer) to account
for things that would not be found by the JVM calculator (if, for
example, there was an external runtime like Tomcat or Liberty). The
value can either be an integer, which will adjust the class count by the
number given, or a string that represents a positive integer ending in
'%'. This will be parsed as an adjustment factor and the class count
will be multiplied by this factor.

Signed-off-by: Josh Ghiloni <jghiloni@vmware.com>